+++ tomcat/src/main/java/org/apache/coyote/http11/SessionManager.java
@@ -23,5 +23,4 @@ public class SessionManager {
     public void putSession(String sessionId, Session session) {
         sessions.put(sessionId, session);
     }
-
 }
