+++ tomcat/src/main/java/org/apache/coyote/http11/response/line/ResponseLine.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.response.line;
 
-import org.apache.coyote.http11.HttpProtocol;
+import org.apache.coyote.http11.common.HttpProtocol;
 
 public class ResponseLine {
